+++ test/functional/postsV2.js
@@ -22,6 +22,8 @@ import {
   hidePost,
   savePost,
   unsavePost,
+  performJSONRequest,
+  authHeaders,
 } from './functional_test_helper';
 
 describe('TimelinesControllerV2', () => {
@@ -308,6 +310,19 @@ describe('TimelinesControllerV2', () => {
         await updatePostAsync(luna, postData);
         const { posts } = await fetchPost(luna.post.id);
         expect(posts.attachments, 'to equal', postData.attachments);
+      });
+
+      it('should not allow to rebind attachment from post1 to new post', async () => {
+        await updatePostAsync(luna, { body: luna.post.body, attachments: [attId1] });
+
+        const resp = await performJSONRequest(
+          'POST',
+          '/v1/posts',
+          { post: { body: 'Body', attachments: [attId1] }, meta: { feeds: [luna.username] } },
+          authHeaders(luna),
+        );
+
+        expect(resp, 'to satisfy', { __httpCode: 403 });
       });
 
       describe('Luna wrote another post', () => {