+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [1.91.0] - Not released
 ### Fixed
 - Properly index texts with HTML-like tags (like spoiler tags)
+- Prevent the theft of attachments from other posts when creating a post.
 
 ### Added
 - The new type of authorization sessions (SessionTokenV1). Sessions can be